add test case
